+++ app/Http/Controllers/api/v1/RoomController.php
@@ -212,7 +212,9 @@ class RoomController extends Controller
                 $name,
                 $room->getRole(Auth::user()),
                 $id,
-                Auth::user() ? Auth::user()->bbb_skip_check_audio : false)
+                Auth::user() ? Auth::user()->bbb_skip_check_audio : false,
+                Auth::user() ? Auth::user()->imageUrl : null
+            )
         ]);
     }
 
@@ -249,7 +251,9 @@ class RoomController extends Controller
                 $name,
                 $room->getRole(Auth::user()),
                 $id,
-                Auth::user() ? Auth::user()->bbb_skip_check_audio : false)
+                Auth::user() ? Auth::user()->bbb_skip_check_audio : false,
+                Auth::user() ? Auth::user()->imageUrl : null
+            )
         ]);
     }