+++ resources/js/lang/de/settings.js
@@ -61,6 +61,20 @@ export default {
     nodata: 'Es sind keine Benutzer vorhanden!',
     nodataFiltered: 'Für die Suchanfrage wurden keine Benutzer gefunden!',
 
+    image: {
+      title: 'Profilbild',
+      upload: 'Neues Bild hochladen',
+      crop: 'Profilbild zuschneiden',
+      save: 'Übernehmen',
+      cancel: 'Abbrechen',
+      delete: 'Bild löschen',
+      undo_delete: 'Löschen rückgänig machen',
+      invalidMime: {
+        message: 'Das Dateiformat wird nicht unterstützt. Bitte wählen Sie eine jpg- oder png-Datei aus.',
+        title: 'Ungültiges Dateiformat'
+      }
+    },
+
     id: 'ID',
     firstname: 'Vorname',
     lastname: 'Nachname',