+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support for parallel testing ([#157], [#158])
 - Modal to confirm end of room membership ([#159], [#165])
+- Profile image for room member list and user avatar in BBB ([#166], [#167])
 
 ### Changed
 - Update BBB api to v4.0.0 ([#155], [#156])
@@ -186,6 +187,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#162]: https://github.com/THM-Health/PILOS/issues/162
 [#163]: https://github.com/THM-Health/PILOS/pull/163
 [#165]: https://github.com/THM-Health/PILOS/pull/165
+[#166]: https://github.com/THM-Health/PILOS/issues/166
+[#167]: https://github.com/THM-Health/PILOS/pull/167
 
 [unreleased]: https://github.com/THM-Health/PILOS/compare/v1.3.0...HEAD
 [1.0.0]: https://github.com/THM-Health/PILOS/releases/tag/v1.0.0